# image-reactor/doc/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Image Processor API v1",
    "summary": "Image metadata processor with FastAPI and asyncpg",
    "description": "Image metadata processor API.\n\n## Features\n- Store and retrieve image metadata\n- Track failed download attempts\n- Health check endpoints\n",
    "contact": {
      "name": "Jan Trojak",
      "email": "jan.trojak@recombee.com"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/failed-downloads-attempts/fail": {
      "post": {
        "tags": [
          "Failed Downloads Attempts"
        ],
        "summary": "Record failed download attempt",
        "description": "Record a failed download attempt. If the image doesn't exist, creates a stub image record first.",
        "operationId": "record_failed_download_failed_downloads_attempts_fail_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailedDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedDownloadAttempt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/failed-downloads-attempts/failed": {
      "get": {
        "tags": [
          "Failed Downloads Attempts"
        ],
        "summary": "List failed download attempts",
        "description": "Retrieve all failed download attempts with optional time period filtering and pagination",
        "operationId": "list_failed_download_attempts_failed_downloads_attempts_failed_get",
        "parameters": [
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (inclusive) for filtering attempts",
              "title": "From Date"
            },
            "description": "Start date (inclusive) for filtering attempts"
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (inclusive) for filtering attempts",
              "title": "To Date"
            },
            "description": "End date (inclusive) for filtering attempts"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of records to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of records to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of records to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FailedDownloadAttempt"
                  },
                  "title": "Response List Failed Download Attempts Failed Downloads Attempts Failed Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/failed-downloads-attempts/{attempt_id}": {
      "get": {
        "tags": [
          "Failed Downloads Attempts"
        ],
        "summary": "Get failed download attempt by ID",
        "description": "Retrieve a specific failed download attempt by its ID",
        "operationId": "get_failed_download_attempt_failed_downloads_attempts__attempt_id__get",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedDownloadAttempt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FailedDownloadAttempt": {
        "properties": {
          "database_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Id",
            "description": "External database identifier"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id",
            "description": "External item identifier"
          },
          "property_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Name",
            "description": "Name of the property"
          },
          "image_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Number",
            "description": "Image number within the property"
          },
          "attempt_status": {
            "type": "string",
            "title": "Attempt Status",
            "description": "Status of the attempt"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "Error message if failed"
          },
          "http_status": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMaximum": 600.0,
                "exclusiveMinimum": 100.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Http Status",
            "description": "HTTP status code"
          },
          "link": {
            "type": "string",
            "title": "Link",
            "description": "Image URL that was attempted"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "attempted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Attempted At"
          },
          "image_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Link",
            "description": "Image link from related image"
          },
          "tries": {
            "type": "integer",
            "title": "Tries",
            "description": "Number of download attempts"
          }
        },
        "type": "object",
        "required": [
          "attempt_status",
          "link",
          "id",
          "attempted_at",
          "tries"
        ],
        "title": "FailedDownloadAttempt",
        "description": "Complete failed download attempt model."
      },
      "FailedDownloadRequest": {
        "properties": {
          "link": {
            "type": "string",
            "title": "Link",
            "description": "Image URL that failed to download"
          },
          "error_message": {
            "type": "string",
            "title": "Error Message",
            "description": "Error message from the failed attempt"
          },
          "http_status": {
            "type": "integer",
            "title": "Http Status",
            "description": "HTTP status code from the failed attempt"
          }
        },
        "type": "object",
        "required": [
          "link",
          "error_message",
          "http_status"
        ],
        "title": "FailedDownloadRequest",
        "description": "Request model for recording a failed download attempt."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}